pyproject.toml
[tool]
[tool.poetry]
name = "pykube-ng"
version = "20.1.0"
description = "Python client library for Kubernetes"
license = "Apache"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3 :: Only",
]
authors = ["Eldarion, Inc. <development@eldarion.com>", "Henning Jacobs <henning@jacobs1.de>"]
readme = "README.rst"
packages = [{ include = "pykube" }]

[tool.poetry.dependencies]
python = ">=3.6"
google-auth = {optional = true, version = "*"}
jsonpath-ng = {optional = true, version = "*"}
pyyaml = "*"
requests = ">=2.12"


[tool.poetry.extras]
gcp = ["google-auth", "jsonpath-ng"]

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.761"
pytest-cov = "^2.8.1"
responses = "^0.10.9"
sphinx = "^2.3.1"
sphinx-rtd-theme = "^0.4.3"
sphinx-autodoc-annotation = "^1.0-1"
flake8 = "^3.7"
pre-commit = "^1.21.0"
